Factor output-supporting code to mpas_base [skip ci]
